+++ .gitignore
@@ -3,3 +3,4 @@ target
 Cargo.lock
 *.swp
 *.dll
+.vscode
